# src/swagger.yaml
openapi: 3.0.0
info:
  title: OKR Express API
  version: 1.0.0
  description: API for managing OKRs, companies, and teams
servers:
  - url: http://localhost:3000
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        email:
          type: string
        companies:
          type: array
          items:
            type: object
            properties:
              _id:
                type: string
              name:
                type: string
        roles:
          type: array
          items:
            type: object
            properties:
              company:
                type: object
                properties:
                  _id:
                    type: string
                  name:
                    type: string
              role:
                type: string
    Team:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        companyId:
          type: string
        members:
          type: array
          items:
            type: object
            properties:
              userId:
                type: string
        requiredRoles:
          type: array
          items:
            type: string
...
paths:
...
  /api/users/{userId}:
    get:
      summary: Get user details
      tags: [Users]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User details including companies and roles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
        '404':
          description: User not found
...
  /api/companies/{companyId}/users/{userId}:
    delete:
      summary: Remove user from company
      tags: [Companies]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: companyId
          required: true
          schema:
            type: string
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User removed from company successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Only company creator can remove users
        '404':
          description: Company not found or User not found
        '400':
          description: User is not in this company
...
  /api/teams/{teamId}/users/{userId}:
    delete:
      summary: Remove user from team
      tags: [Teams]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: teamId
          required: true
          schema:
            type: string
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User removed from team successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  team:
                    $ref: '#/components/schemas/Team'
        '401':
          description: Unauthorized
        '404':
          description: Team not found
        '400':
          description: User is not in this team
...